Add files via upload
A Portuguese bank has observed a decline in its revenue, the primary cause for it being customers not subscribing to long-term deposits. So, the bank wants to run a marketing campaign to invite customers to subscribe to a term deposit. In this project, we predict which customers would subscribe based on the available bank data.